requirements.txt
requests>=2.9.1
beautifulsoup4>=4.4.1
pushbullet.py>=0.10.0